Fix for Android phones
Android phones have a operatin system feature that will display a captive portal shortly after connection to a wifi that requires a log in. This is only triggered if you dont use an adress in the private range!

This is a new request based on
 https://github.com/esp8266/Arduino/pull/5529
and https://github.com/espressif/arduino-esp32/issues/1037

it seems that Android phones will not display the autoredict page if it is in the private IP Range.

I found this fix in the two issues above and tried it with several android phones, non of them displayed it bevore:
Xiaomi Redmi Note 7 (miui 10.3 Android 7) 
Samsung Galaxy A7  (Android 8.0.0) 
Amazon Kindle Fire 7 (Fire OS 5.6.4.0 Android 5.1)

I think for easy access to newcomers this should be changed so it works instantly!